// src/api/generated/models/listAppointmentSlotsApiV1OfficesOfficeIdAppointmentSlotsGetParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentSlotSortField } from './appointmentSlotSortField'
import type { AppointmentSlotStatus } from './appointmentSlotStatus'
import type { SortOrder } from './sortOrder'

export type ListAppointmentSlotsApiV1OfficesOfficeIdAppointmentSlotsGetParams =
  {
    status?: AppointmentSlotStatus | null
    sort_by?: AppointmentSlotSortField
    order?: SortOrder
    starts_from?: string | null
    starts_to?: string | null
    /**
     * @minimum 1
     */
    page?: number
    /**
     * @minimum 1
     * @maximum 100
     */
    size?: number
  }